# openapi/zeus.v1.yaml
openapi: 3.1.0
info:
  title: Zeus API
  description: Enterprise Harness Agent control plane and durable execution API.
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
paths:
...
  /api/v1/openapi.json:
    get:
      tags:
      - platform
      operationId: openapi
      responses:
        '200':
          description: Successful response
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - {}
...
  /api/v1/organizations/{organization_id}/invitations:
    get:
      tags:
      - identity
      operationId: list_invitations
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Organization invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationResponse'
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
    post:
      tags:
      - identity
      operationId: create_invitation
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
        required: true
      responses:
        '201':
          description: Organization invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
  /api/v1/organizations/{organization_id}/invitations/{invitation_id}:
    delete:
      tags:
      - identity
      operationId: revoke_invitation
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: invitation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Invitation revoked
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
  /api/v1/organizations/{organization_id}/invitations/{invitation_id}/resend:
    post:
      tags:
      - identity
      operationId: resend_invitation
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: invitation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: Invitation resend accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedIdentityResponse'
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
...
components:
  schemas:
    AcceptedIdentityResponse:
      type: object
      required:
      - accepted
      - detail
      properties:
        accepted:
          type: boolean
        detail:
          type: string
...
    CreateInvitationRequest:
      type: object
      required:
      - email
      - organization_role
      properties:
        email:
          type: string
        organization_role:
          type: string
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/InvitationWorkspaceRequest'
...
    InvitationResponse:
      type: object
      required:
      - id
      - organization_id
      - invited_by
      - email
      - organization_role
      - status
      - expires_at
      - created_at
      properties:
        accepted_at:
          type:
          - string
          - 'null'
          format: date-time
        accepted_by:
          type:
          - string
          - 'null'
          format: uuid
        created_at:
          type: string
          format: date-time
        email:
          type: string
        expires_at:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        invited_by:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        organization_role:
          type: string
        revoked_at:
          type:
          - string
          - 'null'
          format: date-time
        status:
          type: string
    InvitationWorkspaceRequest:
      type: object
      required:
      - workspace_id
      - role
      properties:
        role:
          type: string
        workspace_id:
          type: string
          format: uuid
...
    ProblemDetails:
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      - request_id
      properties:
        code:
          type: string
        detail:
          type: string
        request_id:
          type: string
          format: uuid
        status:
          type: integer
          format: int32
          minimum: 0
        title:
          type: string
        type:
          type: string
          example: https://zeus.example.com/problems/validation_failed
...
  securitySchemes:
    serviceAccountBearer:
      type: http
      scheme: bearer
      bearerFormat: ServiceAccount
      description: Bearer token for a Zeus service account
    sessionCookie:
      type: apiKey
      in: cookie
      name: zeus_session
      description: HttpOnly session cookie
...